Add support for RelMeAuth
As of this commit, if the user has not added an authorization_endpoint
link to the given profile URL, but links to other social media networks
such as Twitter or GitHub are found tagged as rel="me", DummyAuth will
continue the login process using RelMeAuth.

This means that the user will be redirected to a third party login
provider that can log the user in using third party silo accounts on
Twitter or GitHub.